refactor(login): simplifica a estrutura de arquivos da página de login
